♻️ refactor : ShowDetail 컴포넌트 분리
